tela de edição de perfil e verificação de email adicionados

+++ structure.sql
@@ -2,7 +2,7 @@ create database Fatecarona;
   use Fatecarona;
 
 create table membro(
-  ra int not null primary key,
+  ra bigint not null primary key,
   nome varchar(40) not null,
   email varchar(40) not null,
   genero char(1) default 'M',
